Refactored frontend - moved modals to separate components
