+++ frontend/src/components/SwapComponent/SwapModal.scss
@@ -1,0 +1,78 @@
+.SwapComponent-swap-modal {
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+
+  .SwapComponent-swap-modal-header {
+    display: flex;
+    padding: 10px;
+    font-size: 18px;
+    border-bottom: 1px solid lightgray;
+    font-weight: 700;
+    .SwapComponent-swap-modal-header-image {
+      display: flex;
+      align-items: center;
+      margin-right: 5px;
+    }
+  }
+
+  .SwapComponent-swap-modal-txdetails {
+    display: flex;
+    flex-direction: column;
+    margin-top: 15px;
+
+    .SwapComponent-swap-modal-txdetail {
+      display: flex;
+      flex-direction: row;
+      justify-content: space-between;
+      font-size: 24px;
+      font-weight: 700;
+      padding: 0px 20px;
+    }
+
+    .SwapComponent-arrow {
+      font-size: 16px;
+    }
+
+    .SwapComponent-swap-modal-subtitle {
+      font-size: 12px;
+      font-style: italic;
+      color: gray;
+      font-weight: 700;
+      margin-top: 15px;
+      padding: 15px;
+    }
+
+    .SwapComponent-swap-modal-txinfo {
+      display: flex;
+      justify-content: space-between;
+      font-size: 20px !important;
+      padding: 20px;
+      margin-top: 10px;
+      border: 2px solid lightgray;
+    }
+  }
+
+  .SwapComponent-swap-modal-bottom {
+    margin-top: 25px;
+
+    .SwapComponent-swap-modal-button {
+      background-color: #00c3b4;
+      color: #ffffff;
+      padding: 10px 15px;
+      width: 100%;
+      height: 100%;
+      font-size: 18px;
+      border: 1px solid #00c3b4;
+      border-radius: 12px;
+      text-transform: uppercase;
+      font-weight: 600;
+      cursor: pointer;
+
+      &:hover {
+        background-color: #00ab9e;
+      }
+    }
+  }
+}
